+++ source/tnn/core/default_network.h
@@ -82,7 +82,7 @@ public:
     virtual Status DeInit();
 
     // @brief get network forward for all blob memory size
-    virtual Status GetForwardMemorySize(int &memory_size);
+    virtual Status GetForwardMemorySize(size_t &memory_size);
 
     // @brief set forward memory when share memory mode is set from external
     virtual Status SetForwardMemory(void *memory);
