+++ source/tnn/device/huawei_npu/npu_network.h
@@ -59,7 +59,7 @@ public:
     //  forward
     //  @return error code: If successful, returns zero. Otherwise, returns
     //  an error code.
-    virtual Status GetForwardMemorySize(int &memory_size);
+    virtual Status GetForwardMemorySize(size_t &memory_size);
 
     //  @brief: set memory used by the rapidnet instance without forward
     //  memory, the memory size must be at least that returned by
